Redefined class Domain, and added script for step 2
Class Domain is now more precise (can us .PDB to know the PDB code of a domain, .chain to know it's chain, and .location to get the range of the domain)

The new script (Project_step_2.py) will add to the Domains their pfam accession code. It's really slow, but I'm pretty sure it works.